Implement recursive container deletion (tagging only)

+++ api/dao/basecontainerstorage.py
@@ -45,9 +45,10 @@ class ContainerStorage(object):
     Examples: projects, sessions, acquisitions and collections
     """
 
-    def __init__(self, cont_name, use_object_id=False):
+    def __init__(self, cont_name, use_object_id=False, use_delete_tag=False):
         self.cont_name = cont_name
         self.use_object_id = use_object_id
+        self.use_delete_tag = use_delete_tag
         self.dbc = config.db[cont_name]
 
     @classmethod
@@ -193,6 +194,8 @@ class ContainerStorage(object):
                 _id = bson.ObjectId(_id)
             except bson.errors.InvalidId as e:
                 raise APIStorageException(e.message)
+        if self.use_delete_tag:
+            return self.dbc.update_one({'_id': _id}, {'$set': {'deleted': datetime.datetime.utcnow()}})
         return self.dbc.delete_one({'_id':_id})
 
     def get_el(self, _id, projection=None, fill_defaults=False):
@@ -201,17 +204,20 @@ class ContainerStorage(object):
                 _id = bson.ObjectId(_id)
             except bson.errors.InvalidId as e:
                 raise APIStorageException(e.message)
-        cont = self._from_mongo(self.dbc.find_one(_id, projection))
+        cont = self._from_mongo(self.dbc.find_one({'_id': _id, 'deleted': {'$exists': False}}, projection))
         if fill_defaults:
             cont =  self._fill_default_values(cont)
         return cont
 
     def get_all_el(self, query, user, projection, fill_defaults=False):
+        if query is None:
+            query = {}
         if user:
             if query.get('permissions'):
                 query['$and'] = [{'permissions': {'$elemMatch': user}}, {'permissions': query.pop('permissions')}]
             else:
                 query['permissions'] = {'$elemMatch': user}
+        query['deleted'] = {'$exists': False}
 
         # if projection includes files.info, add new key `info_exists`
         if projection and 'files.info' in projection:
